Javadoc: added helper methods plus examples

+++ commercetools/commercetools-sdk-java-api/src/main/java/com/commercetools/api/models/product/ByIdVariantIdentifier.java
@@ -6,13 +6,13 @@ import com.commercetools.api.models.Referencable;
 public interface ByIdVariantIdentifier extends VariantIdentifier {
     String getProductId();
 
-    Integer getVariantId();
+    Long getVariantId();
 
-    static ByIdVariantIdentifier of(final Referencable<Product> product, final Integer variantId) {
+    static ByIdVariantIdentifier of(final Referencable<Product> product, final Long variantId) {
         return of(product.toReference().getId(), variantId);
     }
 
-    static ByIdVariantIdentifier of(final String productId, final Integer variantId) {
+    static ByIdVariantIdentifier of(final String productId, final Long variantId) {
         return new VariantIdentifierImpl(productId, variantId, null);
     }
 }
